ISE check update
Test existence of ise related directories

+++ scabundle.py
@@ -147,8 +147,10 @@ def connectivity():
 def ona_settings_and_logs():
     copytree("/opt/obsrvbl-ona/", f'{bundledir}/ona_settings/', ignore=ignore_patterns('*python-packages*', '*__pycache__*','*.2*','*pna-*.log*','*pdns_*.pcap.gz'), copy_function=copy)
     copytree("/var/log", f'{bundledir}/var/log/', ignore=ignore_patterns('*.dat','journal'), copy_function=copy)
-    copytree("/etc/ise_poller", f'{bundledir}/etc/ise_poller/', copy_function=copy)
-    copytree("/tmp/ise_poller", f'{bundledir}/tmp/ise_poller/', copy_function=copy)
+    if os.path.exists("/etc/ise_poller"):
+        copytree("/etc/ise_poller", f'{bundledir}/etc/ise_poller/', copy_function=copy)
+    if os.path.exists("/tmp/ise_poller"):
+        copytree("/tmp/ise_poller", f'{bundledir}/tmp/ise_poller/', copy_function=copy)
     cmd = f"/opt/silk/bin/rwcut --timestamp-format iso --fields sIp,dIp,sPort,dPort,protocol,Bytes,Packets,sTime,eTime /opt/obsrvbl-ona/logs/ipfix/.202* >> {bundledir}/ona_settings/logs/ipfix/clear_silk 2>/dev/null"
     run(cmd,shell=True)
